chore: 어드민 App Router 전환 기반 추가
