+++ apps/admin/src/app/providers.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import type { ReactNode } from "react";
+import { Toaster } from "sonner";
+import { QueryProvider } from "@/components/providers/QueryProvider";
+
+interface ProvidersProps {
+	children: ReactNode;
+}
+
+export function Providers({ children }: ProvidersProps) {
+	return (
+		<QueryProvider>
+			{children}
+			<Toaster />
+		</QueryProvider>
+	);
+}
