+++ apps/admin/src/app/scores/page.tsx
@@ -1,0 +1,10 @@
+import { RequireAdminSession } from "@/components/features/auth/RequireAdminSession";
+import { ScoresPageContent } from "@/components/features/scores/ScoresPageContent";
+
+export default function ScoresPage() {
+	return (
+		<RequireAdminSession>
+			<ScoresPageContent />
+		</RequireAdminSession>
+	);
+}
